Add files via upload
This is all of the data science/machine learning I did, the main relevant CSV is allContentMeta, as this is what we read in to then create the other CSV's. Onion Article NLP is preparing the data for the machine learning. Seasoned Onion ML is the training of the machine learning algos.